+++ src/main/java/com/yapp/ios1/service/UserService.java
@@ -42,8 +42,6 @@ public class UserService {
     private final UserMapper userMapper;
     private final PasswordEncoder passwordEncoder;
     private final FollowMapper followMapper;
-    private final NotificationService notificationService;
-    private final AlarmMapper alarmMapper;
 
     @Value("${buok.s3.dir.profile}")
     private String profileDir;
@@ -181,30 +179,5 @@ public class UserService {
     // 친구 리스트
     public List<FriendDto> getFriendList(Long userId) {
         return followMapper.getFollowListByUserId(userId);
-    }
-
-    /**
-     * 팔로우 신청, 알람 로그 저장, 알람 보내기
-     * @param myUserId
-     * @param friendId
-     */
-    @Transactional
-    public void followRequest(Long myUserId, Long friendId) {
-        // Friend INSERT Query 보다 알람을 먼저 보내도록 해놓았음
-        final int FOLLOW_REQUEST = 2;
-        NotificationForOneDto notification = sendFollowAlarmRequest(friendId);
-        followMapper.followRequest(myUserId, friendId, FOLLOW_REQUEST);
-        alarmMapper.insertAlarmLog(notification, friendId);
-    }
-
-    private NotificationForOneDto sendFollowAlarmRequest(Long friendId) {
-        String deviceToken = userMapper.findDeviceTokenByUserId(1L);   // Redis 에서 꺼내오는 걸로 고도화 예정 (1L 는 임시)
-        NotificationForOneDto notificationDto = NotificationForOneDto.builder()
-                .title("팔로우 제목")
-                .message("팔로우 신청 메세지")
-                .deviceToken(deviceToken)
-                .build();
-        notificationService.sendByToken(notificationDto);
-        return notificationDto;
     }
 }
